requirements.txt
numpy~=1.21.2
pdfminer~=20191125
python-slugify~=6.1.2
lxml
beautifulsoup4~=4.10.0
nbformat~=5.1.3
nbconvert~=6.3.0